requirements-base.txt
Django==3.2.12
djangorestframework==3.13.1
django-rest-knox==4.1.0
Pillow==9.0.1
django-imagekit==4.1.0
requests==2.20.1
black==23.1.0

